chore(number-input): move to packages
