+++ packages/components/inputs/number-input/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './number-input';
+
+export { default as version } from './version';
