+++ packages/components/inputs/number-input/src/number-input.visualspec.js
