--- src/components/inputs/number-input/index.js
@@ -1,1 +1,0 @@
-export { default } from './number-input';
